Change plugin export style to named export
