+++ src/language-graphql/index.js
@@ -1,28 +1,8 @@
-import { createRequire } from "node:module";
-import createLanguage from "../utils/create-language.js";
 import printer from "./printer-graphql.js";
-import options from "./options.js";
-import parsers from "./parsers.js";
 
-const require = createRequire(import.meta.url);
-
-const languages = [
-  createLanguage(require("linguist-languages/data/GraphQL.json"), () => ({
-    since: "1.5.0",
-    parsers: ["graphql"],
-    vscodeLanguageIds: ["graphql"],
-  })),
-];
-
-const printers = {
+export const printers = {
   graphql: printer,
 };
-
-const language = {
-  languages,
-  options,
-  printers,
-  parsers,
-};
-
-export default language;
+export { default as languages } from "./languages.evaluate.js";
+export { default as options } from "./options.js";
+export { default as parsers } from "./parsers.js";
