+++ src/language-handlebars/languages.evaluate.js
@@ -1,0 +1,12 @@
+import linguistLanguages from "linguist-languages";
+import createLanguage from "../utils/create-language.js";
+
+const languages = [
+  createLanguage(linguistLanguages.GraphQL, () => ({
+    since: "1.5.0",
+    parsers: ["graphql"],
+    vscodeLanguageIds: ["graphql"],
+  })),
+];
+
+export default languages;
