+++ src/language-markdown/languages.evaluate.js
@@ -1,0 +1,22 @@
+import linguistLanguages from "linguist-languages";
+import createLanguage from "../utils/create-language.js";
+
+const languages = [
+  createLanguage(linguistLanguages.Markdown, (data) => ({
+    since: "1.8.0",
+    parsers: ["markdown"],
+    vscodeLanguageIds: ["markdown"],
+    filenames: [...data.filenames, "README"],
+    extensions: data.extensions.filter((extension) => extension !== ".mdx"),
+  })),
+  createLanguage(linguistLanguages.Markdown, () => ({
+    name: "MDX",
+    since: "1.15.0",
+    parsers: ["mdx"],
+    vscodeLanguageIds: ["mdx"],
+    filenames: [],
+    extensions: [".mdx"],
+  })),
+];
+
+export default languages;
