+++ src/language-yaml/index.js
@@ -1,30 +1,8 @@
-import { createRequire } from "node:module";
-import createLanguage from "../utils/create-language.js";
 import printer from "./printer-yaml.js";
-import options from "./options.js";
-import parsers from "./parsers.js";
 
-const require = createRequire(import.meta.url);
-
-const languages = [
-  createLanguage(require("linguist-languages/data/YAML.json"), (data) => ({
-    since: "1.14.0",
-    parsers: ["yaml"],
-    vscodeLanguageIds: ["yaml", "ansible", "home-assistant"],
-    // yarn.lock is not YAML: https://github.com/yarnpkg/yarn/issues/5629
-    filenames: [
-      ...data.filenames.filter((filename) => filename !== "yarn.lock"),
-      ".prettierrc",
-      ".stylelintrc",
-    ],
-  })),
-];
-
-const language = {
-  languages,
-  printers: { yaml: printer },
-  options,
-  parsers,
+export const printers = {
+  yaml: printer,
 };
-
-export default language;
+export { default as languages } from "./languages.evaluate.js";
+export { default as options } from "./options.js";
+export { default as parsers } from "./parsers.js";
