+++ src/language-yaml/languages.evaluate.js
@@ -1,0 +1,18 @@
+import linguistLanguages from "linguist-languages";
+import createLanguage from "../utils/create-language.js";
+
+const languages = [
+  createLanguage(linguistLanguages.YAML, (data) => ({
+    since: "1.14.0",
+    parsers: ["yaml"],
+    vscodeLanguageIds: ["yaml", "ansible", "home-assistant"],
+    // yarn.lock is not YAML: https://github.com/yarnpkg/yarn/issues/5629
+    filenames: [
+      ...data.filenames.filter((filename) => filename !== "yarn.lock"),
+      ".prettierrc",
+      ".stylelintrc",
+    ],
+  })),
+];
+
+export default languages;
